requirements.txt
pygithub
tqdm

